refactor(types): вынести версионную применимость члена в types.PlatformMemberVersions
Провайдер не должен зависеть от пакета diagnostics. Версионная логика
(targetCompatibilityMode / firesDeprecated / firesUnavailable / sentinel
DEPRECATED_ALWAYS) вынесена из diagnostics.platform.PlatformMemberCalls в
новый types.PlatformMemberVersions — ближе к системе типов. Ею пользуются
и CompletionProvider, и диагностики (DeprecatedMethodCall/UnavailableMemberCall).

PlatformMemberCalls остаётся в diagnostics и отвечает только за резолв
платформенных членов в сайтах вызовов модуля (collect/hasDeletedPrefix) —
это диагностический разбор AST.

Тесты версионной логики вынесены в PlatformMemberVersionsTest; в
PlatformMemberCallsTest остались проверки префикса «Удалить»/«Delete».
