+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/diagnostics/platform/PlatformMemberCalls.java
@@ -21,26 +21,25 @@
  */
 package com.github._1c_syntax.bsl.languageserver.diagnostics.platform;
 
-import com.github._1c_syntax.bsl.languageserver.configuration.LanguageServerConfiguration;
 import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
 import com.github._1c_syntax.bsl.languageserver.types.TypeService;
 import com.github._1c_syntax.bsl.languageserver.types.TypeService.TypedMember;
 import com.github._1c_syntax.bsl.languageserver.types.registry.TypeRegistry;
 import com.github._1c_syntax.bsl.languageserver.utils.Trees;
 import com.github._1c_syntax.bsl.parser.BSLParser;
-import com.github._1c_syntax.bsl.support.CompatibilityMode;
 import org.antlr.v4.runtime.Token;
 import org.eclipse.lsp4j.Position;
 import org.jspecify.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.regex.Pattern;
 
 /**
- * Резолв платформенных членов в сайтах вызовов и сравнение версий — общая
- * база для диагностик устаревания ({@code DeprecatedMethodCall}) и
- * недоступности-по-версии ({@code UnavailableMemberCall}).
+ * Резолв платформенных членов в сайтах вызовов модуля — общая база для
+ * диагностик устаревания ({@code DeprecatedMethodCall}) и
+ * недоступности-по-версии ({@code UnavailableMemberCall}). Версионная
+ * применимость члена (устаревание/недоступность) вынесена в
+ * {@link com.github._1c_syntax.bsl.languageserver.types.PlatformMemberVersions}.
  * <p>
  * Глобальные функции резолвятся напрямую (без инференса), поэтому собираются
  * без pre-filter'а. Члены типов (метод/свойство) предварительно отсеиваются по
@@ -48,19 +47,8 @@ import java.util.regex.Pattern;
  * дешёвый фильтр, после которого {@link TypeService#membersAt} выполняет
  * точный резолв члена на конкретном типе-владельце (иначе сработал бы
  * однофамилец с другого типа).
- * <p>
- * Сравнение версий ведётся относительно режима совместимости проекта. Если он
- * не задан, считается «самая свежая платформа», поэтому любое устаревание
- * срабатывает, а недоступность — никогда.
  */
 public final class PlatformMemberCalls {
-
-  /**
-   * Строка версии вида {@code 8.3.10} — минимум три числовых компонента через
-   * {@code .} или {@code _} (этого достаточно для {@link CompatibilityMode}).
-   * Двухкомпонентные ({@code 8.3}) и прочие — невалидны.
-   */
-  private static final Pattern VERSION_PATTERN = Pattern.compile("^\\d+[._]\\d+[._]\\d+");
 
   /**
    * 1С-конвенция пометки устаревших реквизитов / значений перечислений /
@@ -68,13 +56,6 @@ public final class PlatformMemberCalls {
    * {@code "Delete"} (en-mode).
    */
   private static final String[] DELETED_PREFIXES = {"Удалить", "Delete"};
-
-  /**
-   * Sentinel-значение поля {@code deprecatedSinceVersion} — «устарел всегда»,
-   * безотносительно версии платформы. Используется для oscript-конвенции
-   * (там нет версионирования, но есть API, помеченные устаревшими).
-   */
-  public static final String DEPRECATED_ALWAYS = "*";
 
   private PlatformMemberCalls() {
   }
@@ -168,56 +149,5 @@ public final class PlatformMemberCalls {
     // этого достаточно для membersAt.
     var position = new Position(token.getLine() - 1, token.getCharPositionInLine());
     sink.addAll(typeService.membersAt(documentContext, position));
-  }
-
-  /**
-   * Целевая версия платформы для сравнения. Приоритет: явная настройка
-   * {@code v8platform.targetVersion} в конфиге LS → режим совместимости
-   * конфигурации. Если режим совместимости не задан ({@code DontUse}),
-   * {@link CompatibilityMode} трактует его как самую свежую платформу
-   * (доминирует в {@code compareTo}), поэтому отдельной обработки не требуется.
-   */
-  public static CompatibilityMode targetCompatibilityMode(DocumentContext documentContext,
-                                                          LanguageServerConfiguration configuration) {
-    var explicit = parse(configuration.getV8PlatformOptions().getTargetVersion());
-    if (explicit != null) {
-      return explicit;
-    }
-    return documentContext.getServerContext().getConfiguration().getCompatibilityMode();
-  }
-
-  /**
-   * Член устарел для целевой платформы: {@code target >= deprecatedSinceVersion}.
-   * Sentinel {@link #DEPRECATED_ALWAYS} срабатывает всегда (oscript-конвенция).
-   * Пустая или неразборчивая строка версии → {@code false}.
-   */
-  public static boolean firesDeprecated(String deprecatedSinceVersion, CompatibilityMode target) {
-    if (DEPRECATED_ALWAYS.equals(deprecatedSinceVersion)) {
-      return true;
-    }
-    var version = parse(deprecatedSinceVersion);
-    return version != null && CompatibilityMode.compareTo(version, target) >= 0;
-  }
-
-  /**
-   * Член недоступен в целевой платформе: {@code target < sinceVersion}.
-   * Пустая или неразборчивая строка версии → {@code false}.
-   */
-  public static boolean firesUnavailable(String sinceVersion, CompatibilityMode target) {
-    var version = parse(sinceVersion);
-    return version != null && CompatibilityMode.compareTo(version, target) < 0;
-  }
-
-  /**
-   * Парсит строку версии вида {@code 8.3.10} в {@link CompatibilityMode}.
-   * Пустые, двухкомпонентные ({@code 8.3}) и иные неразборчивые строки →
-   * {@code null} (проверку версии для такого члена пропускаем).
-   */
-  @Nullable
-  private static CompatibilityMode parse(@Nullable String version) {
-    if (version == null || !VERSION_PATTERN.matcher(version).find()) {
-      return null;
-    }
-    return new CompatibilityMode(version);
   }
 }
